pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-research-agent"
dynamic = ["version"]
description = "An intelligent research assistant that generates comprehensive research reports with proper citations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Deep Research Agent Team", email = "your-email@example.com"}
]
maintainers = [
    {name = "Deep Research Agent Team", email = "your-email@example.com"}
]
keywords = ["research", "ai", "citations", "reports", "dspy", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8"
dependencies = [
    "dspy-ai>=2.0.0",
    "google-generativeai>=0.3.0",
    "jinja2>=3.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.0.0",
    "nltk>=3.8.1",
    "tavily-python>=0.1.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]

[project.scripts]
deep-research = "deep_research_agent.cli:main"

[project.urls]
Homepage = "https://github.com/ramsoma/deepresearch"
Documentation = "https://github.com/ramsoma/deepresearch#readme"
Repository = "https://github.com/ramsoma/deepresearch"
"Bug Tracker" = "https://github.com/ramsoma/deepresearch/issues"

[tool.setuptools_scm]
write_to = "deep_research_agent/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["deep_research_agent"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=deep_research_agent",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
] 